fix: 코드 리뷰 반영
- 전역 콜백 cleanup 시 이전 핸들러 복원
- buildWebview에 inputs/outputs 선언으로 증분 빌드 활성화
- 미설치 CLI 타깃 전송 차단
- 동시 스트리밍 시 자동 스크롤 조건을 any streaming으로 확장
- extractObjectText에서 반복 필드 접근을 로컬 변수로 추출
- joinQuietly에서 불필요한 (long) 캐스트 제거
- 클램프/최소값 테스트를 정확값 검증으로 변경
- focus-visible 스타일 추가 및 Firefox 스크롤바 호환성 보강

+++ app/src/main/java/io/github/eottabom/aiagents/providers/CliProcessRunner.java
@@ -380,7 +380,7 @@ final class CliProcessRunner {
 
 	private static void joinQuietly(Thread thread) {
 		try {
-			thread.join((long) 200);
+			thread.join(200);
 		} catch (InterruptedException ex) {
 			Thread.currentThread().interrupt();
 		}
